requirements.txt
numpy
pillow
colorlog